Extracted app logic into GildedRoseApp class while simplifying Program.cs and pointing tests to use GildedRoseApp.
